Move model functions and estimators to own classes
